adding an inital version of the code that allows the beta branch to be built in docker. This allows the building of new worlds on demand and the png of those worlds.:
Fixing the stale map util issue.  It was tied to the fact that we were building
the image and then at the compose level layering on a volume that was removing
the compiled image.  The fix is to take the file at compile time and move it to
a safe location and then move it to the volume at runtime.